Re-sampling to train network. Extract patches for heatmap predictions
Samples, from raw images to train network, are re-generated in a
more exhaustive manner. As a result, far superior predictions are
seen. Further, consecutive patches with stride 3 are generated to
get heatmap predictions. Moreover, morphological operations are
applied to close/fill the holes. Ready to get heatmaps from all the
images and train a classifier from the feaures of the heatmaps.